Time: 0 ms (100%), Space: 23.7 MB (75.67%) - LeetHub

+++ 0637-average-of-levels-in-binary-tree/0637-average-of-levels-in-binary-tree.cpp
@@ -15,7 +15,7 @@ public:
                 if (node->left) q.push(node->left);
                 if (node->right) q.push(node->right);
             }
-            result.push_back(sum / (double)size);
+            result.push_back(sum / size);
         }
         return result;
     }
